Fixed schedule scraping using chromedb
