+++ Dockerfile
@@ -1,8 +1,32 @@
 # Start with the official Golang image
-FROM golang:1.18
+FROM golang:1.23
+
+# Install Chromium and dependencies
+RUN apt-get update && apt-get install -y \
+  chromium \
+  fonts-liberation \
+  libappindicator3-1 \
+  libasound2 \
+  libatk-bridge2.0-0 \
+  libatk1.0-0 \
+  libcups2 \
+  libdbus-1-3 \
+  libgdk-pixbuf2.0-0 \
+  libnspr4 \
+  libnss3 \
+  libxcomposite1 \
+  libxdamage1 \
+  libxrandr2 \
+  xdg-utils \
+  --no-install-recommends && \
+  rm -rf /var/lib/apt/lists/*
 
 # Set the working directory inside the container
 WORKDIR /app
+
+# Set CHROME_PATH so chromedp can find Chromium
+ENV CHROME_PATH=/usr/bin/chromium
+ENV CHROME_BIN=/usr/bin/chromium
 
 # Copy go.mod and go.sum files
 COPY go.mod go.sum ./
