+++ cmd/server/main.go
@@ -6,7 +6,9 @@ import (
 
 	_ "github.com/lib/pq"
 	"github.com/samuel-pratt/bc-ferries-api/cmd/config"
+	"github.com/samuel-pratt/bc-ferries-api/cmd/cron"
 	"github.com/samuel-pratt/bc-ferries-api/cmd/db"
+	"github.com/samuel-pratt/bc-ferries-api/cmd/router"
 )
 
 func main() {
@@ -15,13 +17,13 @@ func main() {
 	db.Init()
 	defer db.Conn.Close()
 
-	SetupCron()
+	cron.SetupCron()
 
 	if config.ServerPort == "" {
 		config.ServerPort = "8080"
 		fmt.Println("INFO: No PORT environment variable detected, defaulting to " + config.ServerPort)
 	}
 
-	router := SetupRouter()
+	router := router.SetupRouter()
 	http.ListenAndServe(":"+config.ServerPort, router)
 }
