+++ go.mod
@@ -1,9 +1,12 @@
 module github.com/samuel-pratt/bc-ferries-api
 
-go 1.18
+go 1.23
+
+toolchain go1.23.10
 
 require (
 	github.com/PuerkitoBio/goquery v1.8.0
+	github.com/chromedp/chromedp v0.13.7
 	github.com/go-co-op/gocron v1.18.0
 	github.com/joho/godotenv v1.5.1
 	github.com/julienschmidt/httprouter v1.3.0
@@ -12,8 +15,15 @@ require (
 
 require (
 	github.com/andybalholm/cascadia v1.3.1 // indirect
+	github.com/chromedp/cdproto v0.0.0-20250403032234-65de8f5d025b // indirect
+	github.com/chromedp/sysutil v1.1.0 // indirect
+	github.com/go-json-experiment/json v0.0.0-20250211171154-1ae217ad3535 // indirect
+	github.com/gobwas/httphead v0.1.0 // indirect
+	github.com/gobwas/pool v0.2.1 // indirect
+	github.com/gobwas/ws v1.4.0 // indirect
 	github.com/robfig/cron/v3 v3.0.1 // indirect
 	github.com/stretchr/testify v1.8.4 // indirect
 	golang.org/x/net v0.19.0 // indirect
 	golang.org/x/sync v0.1.0 // indirect
+	golang.org/x/sys v0.29.0 // indirect
 )
